requirements.txt
tensorflow>=2.8.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
nltk>=3.6.0
jupyter>=1.0.0
ipykernel>=6.0.0
tqdm>=4.62.0
pyyaml>=6.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0